test: split toolbar migration fold suite
